% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_temp.R
\name{gg_temp}
\alias{gg_temp}
\title{Création de graphiques à partir d'un dataframe}
\usage{
gg_temp(data, var_x, var_y)
}
\arguments{
\item{data}{dataframe contenant les données à visualiser}

\item{var_x}{colonne du dataframe à visualiser sur l'axe x}

\item{var_y}{colonne du dataframe à visualiser sur l'axe y}
}
\value{
Un graphique ggplot en points et ligne avec une courbe de tendance uniquement
lorsqu'elle est significative
}
\description{
Création de graphiques à partir d'un dataframe
}
\examples{
\dontrun{g1 <- gg_temp(data = pt_data, var_y = abondance)}
}
